requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.10
dj-database-url==2.1.0
djangorestframework-simplejwt==5.3.0
django-modeltranslation==0.18.11
Pillow==11.3.0
django-storages==1.14.2
drf-spectacular==0.26.5
celery==5.3.4
redis==5.0.1
python-decouple==3.8
django-extensions==3.2.3
django-filter==23.3
sentry-sdk==1.38.0